# src/main/resources/openapi/subtitles-api.yml
openapi: 3.1.0

info:
  title: Video Subtitle API
  description: API para generar subtítulos SRT a partir de un archivo de audio o vídeo.
  version: 1.0.0

servers:
  - url: https://api.example.com
    description: Production

tags:
  - name: Subtitles
    description: Generación de subtítulos

paths:
  /api/v1/subtitles:
    post:
      tags:
        - Subtitles
      summary: Solicitar la generación de subtítulos
      description: >
        Recibe un archivo de audio/vídeo y crea un job de generación de subtítulos.
        El procesamiento es asíncrono: esta petición devuelve inmediatamente un
        identificador de job que debe consultarse en /api/v1/subtitles/{jobId}
        para conocer el estado y, una vez completado, descargar el resultado.
      operationId: createSubtitleJob

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SubtitleRequest'
            encoding:
              file:
                contentType: video/mp4, video/quicktime, video/x-matroska, video/webm, audio/mpeg, audio/mp4, audio/wav

      responses:
        '202':
          description: Job aceptado y encolado para procesamiento.
          headers:
            Location:
              description: URL para consultar el estado del job.
              schema:
                type: string
                example: /api/v1/subtitles/3fa85f64-5717-4562-b3fc-2c963f66afa6
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreatedResponse'

        '400':
          description: Petición inválida (falta el archivo, idioma no soportado, etc).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '413':
          description: El archivo supera el tamaño máximo permitido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '415':
          description: Formato de archivo no soportado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/subtitles/{jobId}:
    get:
      tags:
        - Subtitles
      summary: Consultar el estado de un job de subtítulos
      operationId: getSubtitleJobStatus
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string

      responses:
        '200':
          description: Estado actual del job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'

        '404':
          description: Job no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/subtitles/{jobId}/download:
    get:
      tags:
        - Subtitles
      summary: Descargar el archivo SRT generado
      description: >
        Solo disponible cuando el job está en estado COMPLETED.
      operationId: downloadSubtitleFile
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string

      responses:
        '200':
          description: Archivo SRT generado.
          headers:
            Content-Disposition:
              description: Nombre sugerido para el archivo SRT.
              schema:
                type: string
                example: attachment; filename="subtitles-3fa85f64-5717-4562-b3fc-2c963f66afa6.srt"
          content:
            text/plain:
              schema:
                type: string
                format: binary
              example: |
                1
                00:00:01,000 --> 00:00:04,000
                Hello, how are you?

                2
                00:00:04,500 --> 00:00:07,000
                I'm doing very well, thank you.

        '404':
          description: Job no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '409':
          description: El job todavía no ha finalizado (aún no hay archivo disponible).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:

    SubtitleRequest:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          format: binary
          description: Archivo de audio o vídeo del que se generarán los subtítulos.

        targetLanguage:
          type: string
          description: >
            Código ISO 639-1 del idioma en el que se desean los subtítulos.
            Si no se especifica, se usará el idioma detectado en el propio archivo.
          enum:
            - es
            - en
            - fr
            - de
            - it
            - pt
            - ja
            - ko
            - zh
          nullable: true

    JobCreatedResponse:
      type: object
      required:
        - jobId
        - status
      properties:
        jobId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        status:
          $ref: '#/components/schemas/JobStatus'
        createdAt:
          type: string
          format: date-time

    JobStatusResponse:
      type: object
      required:
        - jobId
        - status
      properties:
        jobId:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        detectedLanguage:
          type: string
          nullable: trueuuid
          description: Idioma detectado en el archivo original (código ISO 639-1).
          example: en
        targetLanguage:
          type: string
          nullable: true
          example: es
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        downloadUrl:
          type: string
          nullable: true
          description: Presente solo cuando status es COMPLETED.
          example: /api/v1/subtitles/3fa85f64-5717-4562-b3fc-2c963f66afa6/download
        error:
          type: string
          nullable: true
          description: Presente solo cuando status es FAILED.

    JobStatus:
      type: string
      enum:
        - PENDING
        - PROCESSING
        - COMPLETED
        - FAILED

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: >
            Código estable del error. Valores usados actualmente por la API:
            JOB_NOT_FOUND, JOB_NOT_READY, UNSUPPORTED_FORMAT, FILE_TOO_LARGE,
            INVALID_REQUEST, INTERNAL_ERROR.
          example: JOB_NOT_READY
        message:
          type: string
          example: El job 3fa85f64-5717-4562-b3fc-2c963f66afa6 no está listo todavía (estado actual PROCESSING)
        details:
          type: string
          nullable: true
          example: mp3, mp4, mkv.